requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
openai==1.6.1
faiss-cpu==1.7.4
scikit-learn==1.3.2
sib-api-v3-sdk==7.6.0
gunicorn==21.2.0